Debian Bookworm: Fixed Bug in dhclient.conf and generation of machine-id
in dhclient.conf the original line:
send host-name = gethostname();

will send the fqdn of the machine, marked as just host-name.

The line is now changed to mark the value correct.
send fqdn.fqdn = gethostname();

In Debian Bookworm, due to a bug in cloud-init 22.4, the machine-id will not be recreated at the first boot.